Time: 2 ms (68.85%), Space: 43.2 MB (6.25%) - LeetHub

+++ 171-excel-sheet-column-number/171-excel-sheet-column-number.java
@@ -8,7 +8,7 @@ class Solution {
             
             int n = columnTitle.charAt(i)-'A'+1;
             // System.out.println(" n = "+n+" j = "+j);
-            count += Math.pow(26, --j)*n;
+            count += (int)Math.pow(26, --j)*n;
         }
         
         return count;
